Moved a lot of files
Basically reorganized the codebase. Mainly around the screens and widgets section.

Also added some listens to CreateTxnForm and CreateTransferTxnForm
