+++ lib/presentation/debug/debug_transactions/debug_transactions_screen.dart
@@ -8,9 +8,9 @@ import 'package:pecunia/features/accounts/domain/entities/account.dart';
 import 'package:pecunia/features/accounts/usecases/get_all_accounts.dart';
 import 'package:pecunia/features/transactions/domain/entities/transaction.dart';
 import 'package:pecunia/features/transactions/usecases/create_transaction.dart';
-import 'package:pecunia/presentation/debug/debug_forms/create_txn_form_widget.dart';
+import 'package:pecunia/presentation/widgets/transactions/forms/create_txn_form_widget.dart';
 import 'package:pecunia/presentation/dialogs/pecunia_dialogs.dart';
-import 'package:pecunia/presentation/screens/shared/recent_txn_list.dart';
+import 'package:pecunia/presentation/widgets/transactions/recent_txn_list.dart';
 
 class DebugTransactionsScreen extends ConsumerWidget {
   const DebugTransactionsScreen({super.key});
