+++ lib/presentation/screens/main_screen.dart
@@ -2,12 +2,12 @@ import 'package:bottom_nav_layout/bottom_nav_layout.dart';
 import 'package:flutter/material.dart';
 import 'package:google_nav_bar/google_nav_bar.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:pecunia/presentation/screens/accounts/accounts_screen.dart';
-import 'package:pecunia/presentation/screens/analytics/analytics_screen.dart';
-import 'package:pecunia/presentation/screens/dashboard/dashboard_screen.dart';
-import 'package:pecunia/presentation/screens/finance/finance_screen.dart';
-import 'package:pecunia/presentation/screens/profile/profile_screen.dart';
-import 'package:pecunia/presentation/screens/shared/pecunia_fab.dart';
+import 'package:pecunia/presentation/screens/primary_screens/accounts_screen.dart';
+import 'package:pecunia/presentation/screens/primary_screens/analytics_screen.dart';
+import 'package:pecunia/presentation/screens/primary_screens/dashboard_screen.dart';
+import 'package:pecunia/presentation/screens/primary_screens/finance_screen.dart';
+import 'package:pecunia/presentation/screens/primary_screens/profile_screen.dart';
+import 'package:pecunia/presentation/widgets/common/pecunia_fab.dart';
 import 'package:simple_animations/simple_animations.dart';
 
 class MainScreen extends HookConsumerWidget {
