+++ lib/presentation/widgets/common/pecunia_fab.dart
